requirements.txt
python-dotenv==1.0.0
redis==5.0.1
asyncpg==0.29.0
aiogram==3.1.1
requests==2.31.0